requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0

pytest==8.0.0
pytest-cov==4.1.0
jsonschema==4.21.1
allure-pytest==2.13.5
pytest-xdist==3.5.0